+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
-## v1.0.0
+## v1.1.0
 
-Initial Release
+- Update Tooling
+- Add label width option
+- Add text size option
+- Truncate labels if too long
+- Sorting bug fix